feat: versioning taskSession messages
