+++ ponos/pkg/executor/signature_validation_test.go
@@ -4,6 +4,10 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"fmt"
+	"math/big"
+	"sync"
+	"testing"
+
 	"github.com/Layr-Labs/crypto-libs/pkg/bn254"
 	ecdsacrypto "github.com/Layr-Labs/crypto-libs/pkg/ecdsa"
 	executorV1 "github.com/Layr-Labs/hourglass-monorepo/ponos/gen/protos/eigenlayer/hourglass/v1/executor"
@@ -22,9 +26,6 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
-	"math/big"
-	"sync"
-	"testing"
 )
 
 // TestValidateTaskSignature_ValidECDSA tests validation of a valid ECDSA signature
@@ -992,7 +993,7 @@ func CreateValidTaskSubmission(t *testing.T, aggregatorKey interface{}, executor
 // SignTaskSubmission signs a task submission with the given key
 func SignTaskSubmission(t *testing.T, task *executorV1.TaskSubmission, signerKey interface{}, executorAddress string) *executorV1.TaskSubmission {
 	// Create the message to sign
-	message := util.EncodeTaskSubmissionMessage(
+	message, _ := util.EncodeTaskSubmissionMessageVersioned(
 		task.TaskId,
 		task.AvsAddress,
 		executorAddress,
@@ -1000,6 +1001,7 @@ func SignTaskSubmission(t *testing.T, task *executorV1.TaskSubmission, signerKey
 		task.TaskBlockNumber,
 		task.OperatorSetId,
 		task.Payload,
+		1,
 	)
 
 	// Sign based on key type
