+++ ponos/protos/eigenlayer/hourglass/v1/executor/executor.proto
@@ -37,7 +37,8 @@ message TaskSubmission {
   uint32 operator_set_id = 6;
   uint32 reference_timestamp = 7;
   string executor_address = 8;
-  uint64 task_block_number = 9;  // Block number for consistent state queries
+  uint64 task_block_number = 9;   // Block number for consistent state queries
+  uint32 version = 10;             // Protocol version (0 = legacy/V1, 2 = V2, etc.)
 }
 
 message TaskResult {
@@ -48,6 +49,7 @@ message TaskResult {
   string avs_address = 5;
   uint32 operator_set_id = 6;
   bytes auth_signature = 7;     // Signs identity data - for authentication
+  uint32 version = 8;           // Protocol version (0 = legacy/V1, 2 = V2, etc.)
 }
 
 // KubernetesConfig contains Kubernetes-specific configuration for a performer
